refactor(scheduler): extract remote dynconfig into separate file
- Move remote dynconfig struct and logic into `dynconfig_remote.go`
- Remove internal-only methods (`GetScheduler`, `GetSeedPeers`,
  `GetSchedulerCluster`, `Get`) from `DynconfigInterface`
- Make them unexported helpers on `remoteDynconfig`
- Add corresponding `dynconfig_remote_test.go` with full test coverage
- Clean up `dynconfig_local.go` to remove stubs for removed interface
  methods
